Fix usage-reporting
Currently Sublime package is trying to execute `$ appbuilder feature-usage-tracking enable`, but this fails as the command had been renamed to usage-reporting. Use correct command and rename all occurrences of feature-usage-tracking to usage-reporting.

Fixes http://teampulse.telerik.com/view#item/296782
